+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * #4455: Qpid Dispatch Router upgraded to 1.12.0
 * #4504: Bump gqlgen from v0.10.2 => v0.11.3
 * #4546: Refactor GraphQL purgeAddress to permit multiple target addresses
+* #4549: Implement GraphQL closeConnections for brokered and standard
 
 ## 0.31.2
 * #4098: Added alert for high broker address memory usage